Fix position encoding for non-equip items
Changed position encoding from writeByte to writeShort for non-equip items to ensure correct data serialization in packet creation.

+++ src/main/java/tools/PacketCreator.java
@@ -418,7 +418,7 @@ public class PacketCreator {
                     p.writeShort(pos);
                 }
             } else {
-                p.writeByte(pos);
+                p.writeShort(pos);
             }
         }
         p.writeByte(itemType);
